Rename and extract the WirelessScan widget to its own file
