+++ test/y2network/widgets/wireless_essid_test.rb
@@ -29,88 +29,3 @@ describe Y2Network::Widgets::WirelessEssid do
 
   include_examples "CWM::CustomWidget"
 end
-
-describe Y2Network::Widgets::WirelessScan do
-  let(:builder) { Y2Network::InterfaceConfigBuilder.for("wlan") }
-  let(:essid) { Y2Network::Widgets::WirelessEssidName.new(builder) }
-  let(:installed) { true }
-  let(:initial_stage) { false }
-  let(:selected_network) do
-    instance_double(Y2Network::WirelessNetwork, essid: "YaST")
-  end
-  let(:wireless_dialog) do
-    instance_double(Y2Network::Dialogs::WirelessNetworks, run: selected_network)
-  end
-
-  subject { described_class.new(builder, update: essid) }
-
-  before do
-    allow(subject).to receive(:scan_supported?).and_return(installed)
-    allow(Yast::Package).to receive(:Installed).and_return(installed)
-    allow(Yast::Stage).to receive(:initial).and_return(initial_stage)
-    allow(essid).to receive(:value=)
-    allow(Y2Network::Dialogs::WirelessNetworks).to receive(:new)
-      .and_return(wireless_dialog)
-  end
-
-  describe "#init" do
-    before do
-      allow(subject).to receive(:present?).and_return(iface_present?)
-    end
-
-    context "when the interface exists" do
-      let(:iface_present?) { true }
-
-      it "does not disable the button" do
-        expect(subject).to_not receive(:disable)
-        subject.init
-      end
-    end
-
-    context "when the interface does not exist" do
-      let(:iface_present?) { false }
-
-      it "disables the button" do
-        expect(subject).to receive(:disable)
-        subject.init
-      end
-    end
-  end
-
-  describe "#handle" do
-    context "when the package for scanning wireless networks is not installed" do
-      let(:installed) { false }
-
-      before do
-        allow(subject).to receive(:scan_supported?).and_call_original
-      end
-
-      it "tries to install it" do
-        expect(Yast::Package).to receive(:Install).and_return(true)
-        subject.handle
-      end
-
-      context "and failed installing the missing package" do
-        it "returns without scanning the available network" do
-          allow(Yast::Package).to receive(:Install).and_return(false)
-          expect(Yast::Popup).to receive(:Error).with(/was not installed/)
-
-          expect(subject.handle).to eql(nil)
-        end
-      end
-    end
-
-    context "when the package for scanning wireless networks is installed" do
-      it "scans the list of available essids" do
-        expect(Y2Network::Dialogs::WirelessNetworks).to receive(:new)
-          .and_return(wireless_dialog)
-        subject.handle
-      end
-
-      it "updates the widget with the list of the available essids with the obtained one" do
-        expect(essid).to receive(:value=).with(selected_network.essid)
-        subject.handle
-      end
-    end
-  end
-end
